requirements.txt
python-dotenv
newsapi-python
google-generativeai